add confirm dialog

+++ imovie-web/imovie-web-ui/src/components/EpisodeListView.vue
@@ -168,7 +168,6 @@ import storageService from '@/services/StorageService'
 import VuePagination from './pagination/VuePagination'
 import VuePaginationInfo from './pagination/VuePaginationInfo'
 import {PaginationEvent} from './pagination/PaginationEvent'
-import $ from 'jquery'
 
 export default {
   name: 'EpisodeListView',
@@ -289,13 +288,15 @@ export default {
       return this.$t('token.updatedTime') + ': ' + updatedTime.split('T')[0]
     },
     deleteMovie: function (id) {
-      movieService.deleteMovie(id, (success, data) => {
-        if (success) {
-          $('div[data-id=' + id + ']').remove()
-        } else {
-          console.log('delete ' + id + ' failed: ' + data)
-        }
-      })
+      this.$dialog.confirm('Are you sure?').then(() => {
+        movieService.deleteMovie(id, (success, data) => {
+          if (success) {
+            this.movies = this.movies.filter(e => e.id !== id)
+          } else {
+            console.log('delete ' + id + ' failed: ' + data)
+          }
+        })
+      }).catch(() => {})
     }
   }
 }
