+++ imovie-web/imovie-web-ui/src/components/FavouriteView.vue
@@ -92,7 +92,6 @@ import userService from '@/services/UserService'
 import VuePagination from './pagination/VuePagination'
 import VuePaginationInfo from './pagination/VuePaginationInfo'
 import {PaginationEvent} from './pagination/PaginationEvent'
-import $ from 'jquery'
 
 export default {
   name: 'FavouriteView',
@@ -196,13 +195,15 @@ export default {
       return this.$t('token.updatedTime') + ': ' + d.getFullYear() + '-' + (d.getMonth() + 1) + '-' + d.getDate()
     },
     deleteFavourite: function (id) {
-      userService.deleteFavourite(id, (success, data) => {
-        if (success) {
-          $('div[data-id=' + id + ']').remove()
-        } else {
-          console.log('delete ' + id + ' failed: ' + data)
-        }
-      })
+      this.$dialog.confirm('Are you sure?').then(() => {
+        userService.deleteFavourite(id, (success, data) => {
+          if (success) {
+            this.movies = this.movies.filter(e => e.id !== id)
+          } else {
+            console.log('delete ' + id + ' failed: ' + data)
+          }
+        })
+      }).catch(() => {})
     }
   }
 }
